~/.xmonad: volume & quodlibet shortcuts.
There's a bug when the master volume is off. Somehow the mute button
doesn't work then which is a little weird. Ideally I'd have the mute
button "pick up" the master volume state and then light up if it is
muted "externally", but it's a bit out of scope for now.