~/.zshrc,~/.bin/up: Fast hg branch detection.
It's quite interesting that even `ruby -e ''` and `python -c ''` take
about 0.03s with >90% system load whereas `git branch` is usually at
0.005s and 0% system load.

And `up some-file` is even faster (0.001s w/ 0% system load), which is
understandable as it mostly does just a few directory hops. But still
quite cool that this now works reliably and fast.

I don't know if it's the gc kicking in or if it just takes that long to
load the standard includes/library, but there's definitively room for
mini-optimization because on slower machines it's too much of an
overhead to use in places like the prompt which has to be responsive.