~/.config/dwm: Add shortcut to put firefox into fullscreen
Useful for reading articles without any distractions, and for watching
video.

This shouldn't be necessary, but seems to be, because of using dwm *and*
Firefox with several addons.

The `windowfocus` command also doesn't seem to work when Firefox already
has focus, this might be another issue with dwm.  Instead, we pass
`--window` to the `key` command directly.