~/.bin/paper: Find papers interactively
This uses `rofi` to display a list of PDFs in a specific directory.  As
the TODOs note, this could be extended to handle other file types,
directories, ...

Also, no need for `tracker` and similar things, which are much more
heavyweight.  But you do need good file names, otherwise this doesn't
very much.