~/.bin/l: Overcomplicate (and view executables)
If the first arg is not a directory or a file, see if we can find it
using `which` and view that file.  E.g. `l l` shows the source code of
`l` itself.

This makes the code much more complicated than before, but I do this so
often, so maybe I'll use it.