~/.config/kanshi: Manage (external) monitors with kanshi
This supports dynamic plug/unplug of monitors, e.g. switching to the
external monitor if it is present, and to the internal display if no
external monitor is connected.

I could not figure out yet how to start up kanshi automatically together
with sway, so right now I only start it when I need it.  (Sway's default
is to use all monitors, which works well.)