.config/{sway/waybar}: Try out waybar's tray and put nm-applet in it
Quite neat and maybe even useful!  The config was a bit odd because
without the tray/icon size the tray icons were just silently invisible.

And the background fix is necessary because otherwise the icons
defaulted to an odd blue background that was a bit distracting.

But it works nicely so far.

See https://github.com/Alexays/Waybar/wiki/Module:-Tray for some more
info.